Move some tests around to the right places
