+++ src/test/java/seedu/duke/userinterface/command/timetable/MarkAsDoneTest.java
@@ -1,4 +1,4 @@
-package seedu.duke.userinterface.command;
+package seedu.duke.userinterface.command.timetable;
 
 import org.junit.jupiter.api.Test;
 import seedu.duke.tasks.TaskList;
